Move down `fopen()` from `Utility` to `Basic` and make it based on AbsolutePath instead of string. Lets us get rid of more `asString` occurrences. Of course, this will later move to `FileSystem` so there is no cleanup of the code here.
