+++ Sources/Utility/Error.swift
@@ -11,10 +11,6 @@
 public enum Error: Swift.Error {
     case obsoleteGitVersion
     case unknownGitError
-    case unicodeDecodingError
-    case unicodeEncodingError
-    case couldNotCreateFile(path: String)
-    case fileDoesNotExist(path: String)
     case invalidPlatformPath
 }
 
@@ -25,10 +21,6 @@ extension Error: CustomStringConvertible {
             return "Git 2.0 or higher is required. Please update git and retry"
         case .unknownGitError:
             return "Failed to invoke git command. Please try updating git"
-        case .unicodeDecodingError: return "Could not decode input file into unicode"
-        case .unicodeEncodingError: return "Could not encode string into unicode"
-        case .couldNotCreateFile(let path): return "Could not create file: \(path)"
-        case .fileDoesNotExist(let path): return "File does not exist: \(path)"
         case .invalidPlatformPath: return "Invalid platform path"
         }
     }
