+++ Sources/Utility/PkgConfig.swift
@@ -126,7 +126,7 @@ struct PkgConfigParser {
             return line
         }
         
-        let file = try fopen(self.pcFile.asString, mode: .read)
+        let file = try fopen(self.pcFile, mode: .read)
         for line in try file.readFileContents().components(separatedBy: "\n") {
             // Remove commented or any trailing comment from the line.
             let uncommentedLine = removeComment(line: line)
