--- Tests/Utility/FileTests.swift
@@ -1,80 +1,0 @@
-/*
- This source file is part of the Swift.org open source project
- 
- Copyright 2015 - 2016 Apple Inc. and the Swift project authors
- Licensed under Apache License v2.0 with Runtime Library Exception
- 
- See http://swift.org/LICENSE.txt for license information
- See http://swift.org/CONTRIBUTORS.txt for Swift project authors
-*/
-
-@testable import Utility
-import XCTest
-
-#if os(Linux)
-    import Foundation  // String.hasSuffix
-#endif
-
-class FileTests: XCTestCase {
-
-    private func loadInputFile(_ name: String) throws -> FileHandle {
-        let input = Path.join(#file, "../Inputs", name).normpath
-        return try fopen(input, mode: .read)
-    }
-    
-    func testOpenFile() {
-        do {
-            let file = try loadInputFile("empty_file")
-            XCTAssertEqual(try file.readFileContents(), "")
-        } catch {
-            XCTFail("The file should be opened without problem")
-        }
-    }
-    
-    func testOpenFileFail() {
-        do {
-            let file = try loadInputFile("file_not_existing")
-            let _ = try file.readFileContents()
-            XCTFail("The file should not be opened since it is not existing")
-        } catch {
-            
-        }
-    }
-    
-    func testReadRegularTextFile() {
-        do {
-            let file = try loadInputFile("regular_text_file")
-            var generator = try file.readFileContents().components(separatedBy: "\n").makeIterator()
-            XCTAssertEqual(generator.next(), "Hello world")
-            XCTAssertEqual(generator.next(), "It is a regular text file.")
-            XCTAssertEqual(generator.next(), "")
-            XCTAssertNil(generator.next())
-        } catch {
-            XCTFail("The file should be opened without problem")
-        }
-    }
-    
-    func testReadRegularTextFileWithSeparator() {
-        do {
-            let file = try loadInputFile("regular_text_file")
-            var generator = try file.readFileContents().components(separatedBy: " ").makeIterator()
-            XCTAssertEqual(generator.next(), "Hello")
-            XCTAssertEqual(generator.next(), "world\nIt")
-            XCTAssertEqual(generator.next(), "is")
-            XCTAssertEqual(generator.next(), "a")
-            XCTAssertEqual(generator.next(), "regular")
-            XCTAssertEqual(generator.next(), "text")
-            XCTAssertEqual(generator.next(), "file.\n")
-            XCTAssertNil(generator.next())
-        } catch {
-            XCTFail("The file should be opened without problem")
-        }
-    }
-
-    static var allTests = [
-        ("testOpenFile", testOpenFile),
-        ("testOpenFileFail", testOpenFileFail),
-        ("testReadRegularTextFile", testReadRegularTextFile),
-        ("testReadRegularTextFileWithSeparator", testReadRegularTextFileWithSeparator)
-    ]
-}
